fix(audit,web): exempt the fleet quota relay and wait for the settings before classifying
Front Desk relays quota snapshots from the primary to every other
member each minute, and every relay was an audit row: 196 of the last
200 on each non-primary member of the live fleet. It is fleet traffic
no operator chose, the same kind as the heartbeat, so it is exempt with
it. The backup classification waits for the settings query so a late
answer does not cost a second preview, and its tests live in their own
file, since the main one had reached the size ceiling.
